backend/requirements.txt
#
# This file is autogenerated by pip-compile with python 3.11
# To update, run:
#
#    pip-compile requirements.in
#
asgiref==3.5.2
    # via django
attrs==22.1.0
    # via jsonschema
cffi==1.15.1
    # via cryptography
colorama==0.4.6
    # via rich
commonmark==0.9.1
    # via rich
cryptography==38.0.3
    # via -r requirements.in
django==4.1.3
    # via
    #   -r requirements.in
    #   django-ckeditor
    #   django-cors-headers
    #   django-filter
    #   django-js-asset
    #   django-rich-logging
    #   django-safedelete
    #   djangorestframework
    #   djangorestframework-simplejwt
    #   drf-spectacular
django-ckeditor==6.5.1
    # via -r requirements.in
django-cors-headers==3.13.0
    # via -r requirements.in
django-filter==22.1
    # via -r requirements.in
django-js-asset==2.0.0
    # via django-ckeditor
django-rich-logging==0.2.0
    # via -r requirements.in
django-safedelete==1.3.1
    # via -r requirements.in
djangorestframework==3.14.0
    # via
    #   -r requirements.in
    #   djangorestframework-simplejwt
    #   drf-spectacular
djangorestframework-simplejwt==5.2.2
    # via -r requirements.in
drf-spectacular==0.24.2
    # via -r requirements.in
drf-writable-nested==0.7.0
    # via -r requirements.in
gunicorn==20.1.0
    # via -r requirements.in
inflection==0.5.1
    # via drf-spectacular
jsonschema==4.17.0
    # via drf-spectacular
pillow==9.3.0
    # via -r requirements.in
pycparser==2.21
    # via cffi
pygments==2.13.0
    # via rich
pyjwt==2.6.0
    # via djangorestframework-simplejwt
pyrsistent==0.19.2
    # via jsonschema
pytz==2022.6
    # via djangorestframework
pyyaml==6.0
    # via drf-spectacular
rich==11.2.0
    # via django-rich-logging
sqlparse==0.4.3
    # via django
uritemplate==4.1.1
    # via drf-spectacular

# The following packages are considered to be unsafe in a requirements file:
# setuptools
